scin_legacy/api/requirements.txt
# Flask API 서비스 의존성
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0

# PyTorch (scin/model/requirements.txt와 동일한 버전)
torch>=2.0.0
torchvision>=0.15.0
efficientnet-pytorch>=0.7.1

# 프로덕션 서버
gunicorn>=21.2.0

# 유틸리티
python-dotenv>=1.0.0
requests>=2.31.0  # 테스트 스크립트용
